+++ resources/views/components/layouts/partials/admin-nav.blade.php
@@ -1,4 +1,0 @@
-{{-- at first glance this feels like a pointless view, however it makes it easier
-to override using ` @includeFirst` when you want to create custom admin navigation --}}
-
-<x-gt-menu menuname="main" filename="nav-admin" withIcons class="menu" />