+++ resources/views/components/layouts/partials/head.blade.php
@@ -1,21 +1,19 @@
-<meta charset="utf-8">
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<meta name="csrf-token" content="{{ csrf_token() }}">
+<meta charset="utf-8" />
+<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-<title>{{ isset($pageTitle) ? Str::title($pageTitle) . ' | ' . config('app.name') : config('app.name') }}</title>
+<title>{{ isset($title) ? Str::title($title) . ' - ' . config('app.name') : config('app.name') }}</title>
 
-<link rel="icon" type="image/x-icon" href="{{ config('gotime.favicon') }}">
-<link href="https://fonts.googleapis.com/css2?family=Nunito:wght@400;600;700&display=swap" rel="stylesheet">
+<link rel="icon" href="/favicon.ico" sizes="any">
+<link rel="icon" href="/favicon.svg" type="image/svg+xml">
+<link rel="apple-touch-icon" href="/apple-touch-icon.png">
+
+<link rel="preconnect" href="https://fonts.googleapis.com">
+<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+<link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap" rel="stylesheet">
 
 @vite(['resources/scss/app.scss', 'resources/js/app.js'])
 
 @livewireStyles
-
-<style>
-    [x-cloak] {
-        display: none !important;
-    }
-</style>
 
 @stack('head')
 @stack('styles')