+++ resources/views/components/layouts/partials/main.blade.php
@@ -1,11 +1,8 @@
+@props(['hasContainer' => false])
 <main {{ $attributes->merge(['class' => 'nk-main']) }}>
 
     @if ($hasContainer)
         <div class="container">
-    @endif
-
-    @if ($hasTitle)
-        <h1>{{ $pageTitle }}</h1>
     @endif
 
     {{ $slot }}